// src/app/shared/models/technology.ts
export enum Technology {
  NextJS = 'NextJS',
  Angular = 'Angular',
  AngularJS = 'AngularJS',
  React = 'React',
  Vue = 'Vue',
  Svelte = 'Svelte',
  NodeJS = 'NodeJS',
  NestJS = 'NestJS',
  Express = 'Express',
  MongoDB = 'MongoDB',
  PostgreSQL = 'PostgreSQL',
  MySQL = 'MySQL',
  TypeScript = 'TypeScript',
  JavaScript = 'JavaScript',
  HTML = 'HTML',
  CSS = 'CSS',
  Sass = 'Sass',
  TailwindCSS = 'TailwindCSS',
  Bootstrap = 'Bootstrap',
  MaterialUI = 'MaterialUI',
  Ionic = 'Ionic',
  Capacitor = 'Capacitor',
  Electron = 'Electron',
  Docker = 'Docker',
  Kubernetes = 'Kubernetes',
  Git = 'Git',
  GitHub = 'GitHub',
  GitLab = 'GitLab',
  Bitbucket = 'Bitbucket',
  Jira = 'Jira',
  Confluence = 'Confluence',
  Slack = 'Slack',
  IntelliJ = 'IntelliJ',
  WebStorm = 'WebStorm',
  VSCode = 'VSCode',
  Vim = 'Vim',
  Windows = 'Windows',
  AWS = 'AWS',
  GCP = 'GCP',
  Azure = 'Azure',
  DigitalOcean = 'DigitalOcean',
  Heroku = 'Heroku',
  Netlify = 'Netlify',
  Vercel = 'Vercel',
  Infomaniak = 'Infomaniak',
  Namecheap = 'Namecheap',
  Cloudflare = 'Cloudflare',
  Figma = 'Figma',
  AdobeXD = 'AdobeXD',
  Photoshop = 'Photoshop',
  Illustrator = 'Illustrator',
  InDesign = 'InDesign',
  PremierePro = 'PremierePro',
  Spring = 'Spring',
  SpringBoot = 'SpringBoot',
  Java = 'Java',
  Kotlin = 'Kotlin',
  Scala = 'Scala',
  Python = 'Python',
  Django = 'Django',
  Flask = 'Flask',
  CSharp = 'CSharp',
  DotNet = 'DotNet',
  PHP = 'PHP',
  Laravel = 'Laravel',
  Grafana = 'Grafana',
  Redis = 'Redis',
  TwitterAPI = 'TwitterAPI',
  Lua = 'Lua',
  FiveM = 'FiveM',
  Stripe = 'Stripe',
}